requirements.txt
ddt
nose
Django==3.0.5
django-bootstrap4
django-filter==2.2.0